+++ src/Server/Search/GetLineExtentCache.cs
@@ -13,7 +13,7 @@ namespace VsChromium.Server.Search {
       _getLineRange = getLineRange;
     }
 
-    public TextRange GetLineExtent(long position) {
+    public TextRange GetLineExtent(int position) {
       if (_previousSpan.HasValue) {
         if (position >= _previousSpan.Value.CharacterOffset &&
             position < _previousSpan.Value.CharacterOffset + _previousSpan.Value.CharacterCount) {
